Improve 2D rendering
